Move tests to root test folder
